[#18] refactor: 소환사 전적 검색 DTO 및 로직 리팩토링
 - DTO 생성 및 적용
 - 데이터 클렌징 class 생성 및 의존성 주입
