+++ src/summoner/summoner.controller.ts
@@ -8,7 +8,7 @@ import {
 } from '@nestjs/common';
 import { ApiOperation, ApiParam, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { HttpExceptionFilter } from 'src/common/exception/http-exception.filter';
-import { SummonerResponseDTO } from './dto/summoner/summoner.dto';
+import { SummonerAllDataDTO } from './dto/summoner/summoner.data.dto';
 import { SummonerService } from './summoner.service';
 
 @Controller('summoner')
@@ -21,7 +21,7 @@ export class SummonerController {
   @ApiResponse({
     status: 200,
     description: '소환사 전적 정보 새로고침',
-    type: SummonerResponseDTO,
+    type: SummonerAllDataDTO,
   })
   @ApiParam({
     name: 'summonerName',
@@ -30,14 +30,13 @@ export class SummonerController {
   })
   @Get('refresh/:summonerName')
   async refreshSummonerInfo(@Param('summonerName') summonerName: string) {
-    return await this.summonerService.RefreshSummoner(summonerName);
+    return await this.summonerService.refreshSummonerData(summonerName);
   }
-
   @ApiOperation({ summary: '소환사 전적 정보 조회' })
   @ApiResponse({
     status: 200,
     description: '소환사 전적 정보 조회',
-    type: SummonerResponseDTO,
+    type: SummonerAllDataDTO,
   })
   @ApiParam({
     name: 'summonerName',
