+++ src/summoner/summoner.module.ts
@@ -2,6 +2,8 @@ import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ChampEntity } from 'src/champ/entities/champ.entity';
+import { SummonerHistoryDataCleansing } from './data-cleansing/history.data.cleansing';
+import { summonerResponseCleansing } from './data-cleansing/summoner.data.cleansing';
 import { SummonerEntity } from './entities/summoner.entity';
 import { SummonerHistoryEntity } from './entities/summoner.history.entity';
 import { SummonerController } from './summoner.controller';
@@ -18,6 +20,11 @@ import { SummonerService } from './summoner.service';
     HttpModule,
   ],
   controllers: [SummonerController],
-  providers: [SummonerService, SummonerRepository],
+  providers: [
+    SummonerService,
+    SummonerRepository,
+    summonerResponseCleansing,
+    SummonerHistoryDataCleansing,
+  ],
 })
 export class SummonerModule {}
